+++ test-utils/database-urls.js
@@ -5,4 +5,5 @@ module.exports = {
   DATABASE_URL_MYSQL_OLD: 'mysql://forest:secret@localhost:8998/lumber-sequelize-test',
   DATABASE_URL_MYSQL: 'mysql://forest:secret@localhost:8999/lumber-sequelize-test',
   DATABASE_URL_POSTGRESQL: 'postgres://forest:secret@localhost:54369/lumber-sequelize-test',
+  DATABASE_URL_POSTGRESQL_OLD: 'postgres://forest:secret@localhost:54368/lumber-sequelize-test',
 };